move file to submitted folder
